+++ packages/rest/src/lib/REST.ts
@@ -1,288 +1,155 @@
-import type { Readable } from 'node:stream';
-import type { ReadableStream } from 'node:stream/web';
-import type { Collection } from '@discordjs/collection';
+import { Collection } from '@discordjs/collection';
+import { DiscordSnowflake } from '@sapphire/snowflake';
 import { AsyncEventEmitter } from '@vladfrangu/async_event_emitter';
-import type { Dispatcher, RequestInit, Response } from 'undici';
+import { filetypeinfo } from 'magic-bytes.js';
+import type { RequestInit, BodyInit, Dispatcher } from 'undici';
 import { CDN } from './CDN.js';
-import {
-	RequestManager,
-	RequestMethod,
-	type HashData,
-	type HandlerRequestData,
-	type InternalRequest,
-	type RequestData,
-	type RouteLike,
-} from './RequestManager.js';
+import { BurstHandler } from './handlers/BurstHandler.js';
+import { SequentialHandler } from './handlers/SequentialHandler.js';
 import type { IHandler } from './interfaces/Handler.js';
-import { DefaultRestOptions, RESTEvents } from './utils/constants.js';
-import { parseResponse } from './utils/utils.js';
+import {
+	BurstHandlerMajorIdKey,
+	DefaultRestOptions,
+	DefaultUserAgent,
+	OverwrittenMimeTypes,
+	RESTEvents,
+} from './utils/constants.js';
+import { RequestMethod } from './utils/types.js';
+import type {
+	RESTOptions,
+	ResponseLike,
+	RestEventsMap,
+	HashData,
+	InternalRequest,
+	RouteLike,
+	RequestHeaders,
+	RouteData,
+	RequestData,
+} from './utils/types.js';
+import { isBufferLike, parseResponse } from './utils/utils.js';
 
 /**
- * Options to be passed when creating the REST instance
+ * Represents the class that manages handlers for endpoints
  */
-export interface RESTOptions {
-	/**
-	 * The agent to set globally
-	 */
-	agent: Dispatcher | null;
-	/**
-	 * The base api path, without version
-	 *
-	 * @defaultValue `'https://discord.com/api'`
-	 */
-	api: string;
-	/**
-	 * The authorization prefix to use for requests, useful if you want to use
-	 * bearer tokens
-	 *
-	 * @defaultValue `'Bot'`
-	 */
-	authPrefix: 'Bearer' | 'Bot';
-	/**
-	 * The cdn path
-	 *
-	 * @defaultValue `'https://cdn.discordapp.com'`
-	 */
-	cdn: string;
-	/**
-	 * How many requests to allow sending per second (Infinity for unlimited, 50 for the standard global limit used by Discord)
-	 *
-	 * @defaultValue `50`
-	 */
-	globalRequestsPerSecond: number;
-	/**
-	 * The amount of time in milliseconds that passes between each hash sweep. (defaults to 1h)
-	 *
-	 * @defaultValue `3_600_000`
-	 */
-	handlerSweepInterval: number;
-	/**
-	 * The maximum amount of time a hash can exist in milliseconds without being hit with a request (defaults to 24h)
-	 *
-	 * @defaultValue `86_400_000`
-	 */
-	hashLifetime: number;
-	/**
-	 * The amount of time in milliseconds that passes between each hash sweep. (defaults to 4h)
-	 *
-	 * @defaultValue `14_400_000`
-	 */
-	hashSweepInterval: number;
-	/**
-	 * Additional headers to send for all API requests
-	 *
-	 * @defaultValue `{}`
-	 */
-	headers: Record<string, string>;
-	/**
-	 * The number of invalid REST requests (those that return 401, 403, or 429) in a 10 minute window between emitted warnings (0 for no warnings).
-	 * That is, if set to 500, warnings will be emitted at invalid request number 500, 1000, 1500, and so on.
-	 *
-	 * @defaultValue `0`
-	 */
-	invalidRequestWarningInterval: number;
-	/**
-	 * The method called to perform the actual HTTP request given a url and web `fetch` options
-	 * For example, to use global fetch, simply provide `makeRequest: fetch`
-	 *
-	 * @defaultValue `undici.request`
-	 */
-	makeRequest(url: string, init: RequestInit): Promise<ResponseLike>;
-	/**
-	 * The extra offset to add to rate limits in milliseconds
-	 *
-	 * @defaultValue `50`
-	 */
-	offset: number;
-	/**
-	 * Determines how rate limiting and pre-emptive throttling should be handled.
-	 * When an array of strings, each element is treated as a prefix for the request route
-	 * (e.g. `/channels` to match any route starting with `/channels` such as `/channels/:id/messages`)
-	 * for which to throw {@link RateLimitError}s. All other request routes will be queued normally
-	 *
-	 * @defaultValue `null`
-	 */
-	rejectOnRateLimit: RateLimitQueueFilter | string[] | null;
-	/**
-	 * The number of retries for errors with the 500 code, or errors
-	 * that timeout
-	 *
-	 * @defaultValue `3`
-	 */
-	retries: number;
-	/**
-	 * The time to wait in milliseconds before a request is aborted
-	 *
-	 * @defaultValue `15_000`
-	 */
-	timeout: number;
-	/**
-	 * Extra information to add to the user agent
-	 *
-	 * @defaultValue DefaultUserAgentAppendix
-	 */
-	userAgentAppendix: string;
-	/**
-	 * The version of the API to use
-	 *
-	 * @defaultValue `'10'`
-	 */
-	version: string;
-}
-
-/**
- * Data emitted on `RESTEvents.RateLimited`
- */
-export interface RateLimitData {
-	/**
-	 * Whether the rate limit that was reached was the global limit
-	 */
-	global: boolean;
-	/**
-	 * The bucket hash for this request
-	 */
-	hash: string;
-	/**
-	 * The amount of requests we can perform before locking requests
-	 */
-	limit: number;
-	/**
-	 * The major parameter of the route
-	 *
-	 * For example, in `/channels/x`, this will be `x`.
-	 * If there is no major parameter (e.g: `/bot/gateway`) this will be `global`.
-	 */
-	majorParameter: string;
-	/**
-	 * The HTTP method being performed
-	 */
-	method: string;
-	/**
-	 * The route being hit in this request
-	 */
-	route: string;
-	/**
-	 * The time, in milliseconds, until the request-lock is reset
-	 */
-	timeToReset: number;
-	/**
-	 * The full URL for this request
-	 */
-	url: string;
-}
-
-/**
- * A function that determines whether the rate limit hit should throw an Error
- */
-export type RateLimitQueueFilter = (rateLimitData: RateLimitData) => Promise<boolean> | boolean;
-
-export interface APIRequest {
-	/**
-	 * The data that was used to form the body of this request
-	 */
-	data: HandlerRequestData;
-	/**
-	 * The HTTP method used in this request
-	 */
-	method: string;
-	/**
-	 * Additional HTTP options for this request
-	 */
-	options: RequestInit;
-	/**
-	 * The full path used to make the request
-	 */
-	path: RouteLike;
-	/**
-	 * The number of times this request has been attempted
-	 */
-	retries: number;
-	/**
-	 * The API route identifying the ratelimit for this request
-	 */
-	route: string;
-}
-
-export interface ResponseLike
-	extends Pick<Response, 'arrayBuffer' | 'bodyUsed' | 'headers' | 'json' | 'ok' | 'status' | 'statusText' | 'text'> {
-	body: Readable | ReadableStream | null;
-}
-
-export interface InvalidRequestWarningData {
-	/**
-	 * Number of invalid requests that have been made in the window
-	 */
-	count: number;
-	/**
-	 * Time in milliseconds remaining before the count resets
-	 */
-	remainingTime: number;
-}
-
-export interface RestEvents {
-	handlerSweep: [sweptHandlers: Collection<string, IHandler>];
-	hashSweep: [sweptHashes: Collection<string, HashData>];
-	invalidRequestWarning: [invalidRequestInfo: InvalidRequestWarningData];
-	rateLimited: [rateLimitInfo: RateLimitData];
-	response: [request: APIRequest, response: ResponseLike];
-	restDebug: [info: string];
-}
-
-export type RestEventsMap = {
-	[K in keyof RestEvents]: RestEvents[K];
-};
-
 export class REST extends AsyncEventEmitter<RestEventsMap> {
+	/**
+	 * The {@link https://undici.nodejs.org/#/docs/api/Agent | Agent} for all requests
+	 * performed by this manager.
+	 */
+	public agent: Dispatcher | null = null;
+
 	public readonly cdn: CDN;
 
-	public readonly requestManager: RequestManager;
+	/**
+	 * The number of requests remaining in the global bucket
+	 */
+	public globalRemaining: number;
+
+	/**
+	 * The promise used to wait out the global rate limit
+	 */
+	public globalDelay: Promise<void> | null = null;
+
+	/**
+	 * The timestamp at which the global bucket resets
+	 */
+	public globalReset = -1;
+
+	/**
+	 * API bucket hashes that are cached from provided routes
+	 */
+	public readonly hashes = new Collection<string, HashData>();
+
+	/**
+	 * Request handlers created from the bucket hash and the major parameters
+	 */
+	public readonly handlers = new Collection<string, IHandler>();
+
+	#token: string | null = null;
+
+	private hashTimer!: NodeJS.Timer | number;
+
+	private handlerTimer!: NodeJS.Timer | number;
+
+	public readonly options: RESTOptions;
 
 	public constructor(options: Partial<RESTOptions> = {}) {
 		super();
 		this.cdn = new CDN(options.cdn ?? DefaultRestOptions.cdn);
-		this.requestManager = new RequestManager(options)
-			// @ts-expect-error For some reason ts can't infer these types
-			.on(RESTEvents.Debug, this.emit.bind(this, RESTEvents.Debug))
-			// @ts-expect-error For some reason ts can't infer these types
-			.on(RESTEvents.RateLimited, this.emit.bind(this, RESTEvents.RateLimited))
-			// @ts-expect-error For some reason ts can't infer these types
-			.on(RESTEvents.InvalidRequestWarning, this.emit.bind(this, RESTEvents.InvalidRequestWarning))
-			// @ts-expect-error For some reason ts can't infer these types
-			.on(RESTEvents.HashSweep, this.emit.bind(this, RESTEvents.HashSweep));
+		this.options = { ...DefaultRestOptions, ...options };
+		this.options.offset = Math.max(0, this.options.offset);
+		this.globalRemaining = Math.max(1, this.options.globalRequestsPerSecond);
+		this.agent = options.agent ?? null;
 
-		this.on('newListener', (name, listener) => {
-			if (name === RESTEvents.Response) this.requestManager.on(name, listener);
-		});
-		this.on('removeListener', (name, listener) => {
-			if (name === RESTEvents.Response) this.requestManager.off(name, listener);
-		});
+		// Start sweepers
+		this.setupSweepers();
 	}
 
-	/**
-	 * Gets the agent set for this instance
-	 */
-	public getAgent() {
-		return this.requestManager.agent;
-	}
+	private setupSweepers() {
+		// eslint-disable-next-line unicorn/consistent-function-scoping
+		const validateMaxInterval = (interval: number) => {
+			if (interval > 14_400_000) {
+				throw new Error('Cannot set an interval greater than 4 hours');
+			}
+		};
 
-	/**
-	 * Sets the default agent to use for requests performed by this instance
-	 *
-	 * @param agent - Sets the agent to use
-	 */
-	public setAgent(agent: Dispatcher) {
-		this.requestManager.setAgent(agent);
-		return this;
-	}
+		if (this.options.hashSweepInterval !== 0 && this.options.hashSweepInterval !== Number.POSITIVE_INFINITY) {
+			validateMaxInterval(this.options.hashSweepInterval);
+			this.hashTimer = setInterval(() => {
+				const sweptHashes = new Collection<string, HashData>();
+				const currentDate = Date.now();
 
-	/**
-	 * Sets the authorization token that should be used for requests
-	 *
-	 * @param token - The authorization token to use
-	 */
-	public setToken(token: string) {
-		this.requestManager.setToken(token);
-		return this;
+				// Begin sweeping hash based on lifetimes
+				this.hashes.sweep((val, key) => {
+					// `-1` indicates a global hash
+					if (val.lastAccess === -1) return false;
+
+					// Check if lifetime has been exceeded
+					const shouldSweep = Math.floor(currentDate - val.lastAccess) > this.options.hashLifetime;
+
+					// Add hash to collection of swept hashes
+					if (shouldSweep) {
+						// Add to swept hashes
+						sweptHashes.set(key, val);
+
+						// Emit debug information
+						this.emit(RESTEvents.Debug, `Hash ${val.value} for ${key} swept due to lifetime being exceeded`);
+					}
+
+					return shouldSweep;
+				});
+
+				// Fire event
+				this.emit(RESTEvents.HashSweep, sweptHashes);
+			}, this.options.hashSweepInterval);
+
+			this.hashTimer.unref?.();
+		}
+
+		if (this.options.handlerSweepInterval !== 0 && this.options.handlerSweepInterval !== Number.POSITIVE_INFINITY) {
+			validateMaxInterval(this.options.handlerSweepInterval);
+			this.handlerTimer = setInterval(() => {
+				const sweptHandlers = new Collection<string, IHandler>();
+
+				// Begin sweeping handlers based on activity
+				this.handlers.sweep((val, key) => {
+					const { inactive } = val;
+
+					// Collect inactive handlers
+					if (inactive) {
+						sweptHandlers.set(key, val);
+						this.emit(RESTEvents.Debug, `Handler ${val.id} for ${key} swept due to being inactive`);
+					}
+
+					return inactive;
+				});
+
+				// Fire event
+				this.emit(RESTEvents.HandlerSweep, sweptHandlers);
+			}, this.options.handlerSweepInterval);
+
+			this.handlerTimer.unref?.();
+		}
 	}
 
 	/**
@@ -341,16 +208,259 @@ export class REST extends AsyncEventEmitter<RestEventsMap> {
 	 * @param options - Request options
 	 */
 	public async request(options: InternalRequest) {
-		const response = await this.raw(options);
+		const response = await this.queueRequest(options);
 		return parseResponse(response);
 	}
 
 	/**
-	 * Runs a request from the API, yielding the raw Response object
+	 * Sets the default agent to use for requests performed by this manager
 	 *
-	 * @param options - Request options
+	 * @param agent - The agent to use
 	 */
-	public async raw(options: InternalRequest) {
-		return this.requestManager.queueRequest(options);
+	public setAgent(agent: Dispatcher) {
+		this.agent = agent;
+		return this;
+	}
+
+	/**
+	 * Sets the authorization token that should be used for requests
+	 *
+	 * @param token - The authorization token to use
+	 */
+	public setToken(token: string) {
+		this.#token = token;
+		return this;
+	}
+
+	/**
+	 * Queues a request to be sent
+	 *
+	 * @param request - All the information needed to make a request
+	 * @returns The response from the api request
+	 */
+	public async queueRequest(request: InternalRequest): Promise<ResponseLike> {
+		// Generalize the endpoint to its route data
+		const routeId = REST.generateRouteData(request.fullRoute, request.method);
+		// Get the bucket hash for the generic route, or point to a global route otherwise
+		const hash = this.hashes.get(`${request.method}:${routeId.bucketRoute}`) ?? {
+			value: `Global(${request.method}:${routeId.bucketRoute})`,
+			lastAccess: -1,
+		};
+
+		// Get the request handler for the obtained hash, with its major parameter
+		const handler =
+			this.handlers.get(`${hash.value}:${routeId.majorParameter}`) ??
+			this.createHandler(hash.value, routeId.majorParameter);
+
+		// Resolve the request into usable fetch options
+		const { url, fetchOptions } = await this.resolveRequest(request);
+
+		// Queue the request
+		return handler.queueRequest(routeId, url, fetchOptions, {
+			body: request.body,
+			files: request.files,
+			auth: request.auth !== false,
+			signal: request.signal,
+		});
+	}
+
+	/**
+	 * Creates a new rate limit handler from a hash, based on the hash and the major parameter
+	 *
+	 * @param hash - The hash for the route
+	 * @param majorParameter - The major parameter for this handler
+	 * @internal
+	 */
+	private createHandler(hash: string, majorParameter: string) {
+		// Create the async request queue to handle requests
+		const queue =
+			majorParameter === BurstHandlerMajorIdKey
+				? new BurstHandler(this, hash, majorParameter)
+				: new SequentialHandler(this, hash, majorParameter);
+		// Save the queue based on its id
+		this.handlers.set(queue.id, queue);
+
+		return queue;
+	}
+
+	/**
+	 * Formats the request data to a usable format for fetch
+	 *
+	 * @param request - The request data
+	 */
+	private async resolveRequest(request: InternalRequest): Promise<{ fetchOptions: RequestInit; url: string }> {
+		const { options } = this;
+
+		let query = '';
+
+		// If a query option is passed, use it
+		if (request.query) {
+			const resolvedQuery = request.query.toString();
+			if (resolvedQuery !== '') {
+				query = `?${resolvedQuery}`;
+			}
+		}
+
+		// Create the required headers
+		const headers: RequestHeaders = {
+			...this.options.headers,
+			'User-Agent': `${DefaultUserAgent} ${options.userAgentAppendix}`.trim(),
+		};
+
+		// If this request requires authorization (allowing non-"authorized" requests for webhooks)
+		if (request.auth !== false) {
+			// If we haven't received a token, throw an error
+			if (!this.#token) {
+				throw new Error('Expected token to be set for this request, but none was present');
+			}
+
+			headers.Authorization = `${request.authPrefix ?? this.options.authPrefix} ${this.#token}`;
+		}
+
+		// If a reason was set, set it's appropriate header
+		if (request.reason?.length) {
+			headers['X-Audit-Log-Reason'] = encodeURIComponent(request.reason);
+		}
+
+		// Format the full request URL (api base, optional version, endpoint, optional querystring)
+		const url = `${options.api}${request.versioned === false ? '' : `/v${options.version}`}${
+			request.fullRoute
+		}${query}`;
+
+		let finalBody: RequestInit['body'];
+		let additionalHeaders: Record<string, string> = {};
+
+		if (request.files?.length) {
+			const formData = new FormData();
+
+			// Attach all files to the request
+			for (const [index, file] of request.files.entries()) {
+				const fileKey = file.key ?? `files[${index}]`;
+
+				// https://developer.mozilla.org/en-US/docs/Web/API/FormData/append#parameters
+				// FormData.append only accepts a string or Blob.
+				// https://developer.mozilla.org/en-US/docs/Web/API/Blob/Blob#parameters
+				// The Blob constructor accepts TypedArray/ArrayBuffer, strings, and Blobs.
+				if (isBufferLike(file.data)) {
+					// Try to infer the content type from the buffer if one isn't passed
+					let contentType = file.contentType;
+
+					if (!contentType) {
+						const [parsedType] = filetypeinfo(file.data);
+
+						if (parsedType) {
+							contentType =
+								OverwrittenMimeTypes[parsedType.mime as keyof typeof OverwrittenMimeTypes] ??
+								parsedType.mime ??
+								'application/octet-stream';
+						}
+					}
+
+					formData.append(fileKey, new Blob([file.data], { type: contentType }), file.name);
+				} else {
+					formData.append(fileKey, new Blob([`${file.data}`], { type: file.contentType }), file.name);
+				}
+			}
+
+			// If a JSON body was added as well, attach it to the form data, using payload_json unless otherwise specified
+			// eslint-disable-next-line no-eq-null, eqeqeq
+			if (request.body != null) {
+				if (request.appendToFormData) {
+					for (const [key, value] of Object.entries(request.body as Record<string, unknown>)) {
+						formData.append(key, value);
+					}
+				} else {
+					formData.append('payload_json', JSON.stringify(request.body));
+				}
+			}
+
+			// Set the final body to the form data
+			finalBody = formData;
+
+			// eslint-disable-next-line no-eq-null, eqeqeq
+		} else if (request.body != null) {
+			if (request.passThroughBody) {
+				finalBody = request.body as BodyInit;
+			} else {
+				// Stringify the JSON data
+				finalBody = JSON.stringify(request.body);
+				// Set the additional headers to specify the content-type
+				additionalHeaders = { 'Content-Type': 'application/json' };
+			}
+		}
+
+		const method = request.method.toUpperCase();
+
+		// The non null assertions in the following block are due to exactOptionalPropertyTypes, they have been tested to work with undefined
+		const fetchOptions: RequestInit = {
+			// Set body to null on get / head requests. This does not follow fetch spec (likely because it causes subtle bugs) but is aligned with what request was doing
+			body: ['GET', 'HEAD'].includes(method) ? null : finalBody!,
+			headers: { ...request.headers, ...additionalHeaders, ...headers } as Record<string, string>,
+			method,
+			// Prioritize setting an agent per request, use the agent for this instance otherwise.
+			dispatcher: request.dispatcher ?? this.agent ?? undefined!,
+		};
+
+		return { url, fetchOptions };
+	}
+
+	/**
+	 * Stops the hash sweeping interval
+	 */
+	public clearHashSweeper() {
+		clearInterval(this.hashTimer);
+	}
+
+	/**
+	 * Stops the request handler sweeping interval
+	 */
+	public clearHandlerSweeper() {
+		clearInterval(this.handlerTimer);
+	}
+
+	/**
+	 * Generates route data for an endpoint:method
+	 *
+	 * @param endpoint - The raw endpoint to generalize
+	 * @param method - The HTTP method this endpoint is called without
+	 * @internal
+	 */
+	private static generateRouteData(endpoint: RouteLike, method: RequestMethod): RouteData {
+		if (endpoint.startsWith('/interactions/') && endpoint.endsWith('/callback')) {
+			return {
+				majorParameter: BurstHandlerMajorIdKey,
+				bucketRoute: '/interactions/:id/:token/callback',
+				original: endpoint,
+			};
+		}
+
+		const majorIdMatch = /^\/(?:channels|guilds|webhooks)\/(\d{17,19})/.exec(endpoint);
+
+		// Get the major id for this route - global otherwise
+		const majorId = majorIdMatch?.[1] ?? 'global';
+
+		const baseRoute = endpoint
+			// Strip out all ids
+			.replaceAll(/\d{17,19}/g, ':id')
+			// Strip out reaction as they fall under the same bucket
+			.replace(/\/reactions\/(.*)/, '/reactions/:reaction');
+
+		let exceptions = '';
+
+		// Hard-Code Old Message Deletion Exception (2 week+ old messages are a different bucket)
+		// https://github.com/discord/discord-api-docs/issues/1295
+		if (method === RequestMethod.Delete && baseRoute === '/channels/:id/messages/:id') {
+			const id = /\d{17,19}$/.exec(endpoint)![0]!;
+			const timestamp = DiscordSnowflake.timestampFrom(id);
+			if (Date.now() - timestamp > 1_000 * 60 * 60 * 24 * 14) {
+				exceptions += '/Delete Old Message';
+			}
+		}
+
+		return {
+			majorParameter: majorId,
+			bucketRoute: baseRoute + exceptions,
+			original: endpoint,
+		};
 	}
 }
