+++ packages/rest/src/lib/handlers/BurstHandler.ts
@@ -1,8 +1,8 @@
 import type { RequestInit } from 'undici';
-import type { ResponseLike } from '../REST.js';
-import type { HandlerRequestData, RequestManager, RouteData } from '../RequestManager.js';
+import type { REST } from '../REST.js';
 import type { IHandler } from '../interfaces/Handler.js';
 import { RESTEvents } from '../utils/constants.js';
+import type { ResponseLike, HandlerRequestData, RouteData } from '../utils/types.js';
 import { onRateLimit, sleep } from '../utils/utils.js';
 import { handleErrors, incrementInvalidCount, makeNetworkRequest } from './Shared.js';
 
@@ -31,7 +31,7 @@ export class BurstHandler implements IHandler {
 	 * @param majorParameter - The major parameter for this handler
 	 */
 	public constructor(
-		private readonly manager: RequestManager,
+		private readonly manager: REST,
 		private readonly hash: string,
 		private readonly majorParameter: string,
 	) {
