+++ packages/rest/src/lib/handlers/Shared.ts
@@ -1,10 +1,10 @@
 import type { RequestInit } from 'undici';
-import type { ResponseLike } from '../REST.js';
-import type { HandlerRequestData, RequestManager, RouteData } from '../RequestManager.js';
+import type { REST } from '../REST.js';
 import type { DiscordErrorData, OAuthErrorData } from '../errors/DiscordAPIError.js';
 import { DiscordAPIError } from '../errors/DiscordAPIError.js';
 import { HTTPError } from '../errors/HTTPError.js';
 import { RESTEvents } from '../utils/constants.js';
+import type { ResponseLike, HandlerRequestData, RouteData } from '../utils/types.js';
 import { parseResponse, shouldRetry } from '../utils/utils.js';
 
 /**
@@ -22,7 +22,7 @@ let invalidCountResetTime: number | null = null;
  *
  * @internal
  */
-export function incrementInvalidCount(manager: RequestManager) {
+export function incrementInvalidCount(manager: REST) {
 	if (!invalidCountResetTime || invalidCountResetTime < Date.now()) {
 		invalidCountResetTime = Date.now() + 1_000 * 60 * 10;
 		invalidCount = 0;
@@ -55,7 +55,7 @@ export function incrementInvalidCount(manager: RequestManager) {
  * @internal
  */
 export async function makeNetworkRequest(
-	manager: RequestManager,
+	manager: REST,
 	routeId: RouteData,
 	url: string,
 	options: RequestInit,
@@ -118,7 +118,7 @@ export async function makeNetworkRequest(
  * @returns - The response if the status code is not handled or null to request a retry
  */
 export async function handleErrors(
-	manager: RequestManager,
+	manager: REST,
 	res: ResponseLike,
 	method: string,
 	url: string,
