+++ packages/rest/src/shared.ts
@@ -2,9 +2,9 @@ export * from './lib/CDN.js';
 export * from './lib/errors/DiscordAPIError.js';
 export * from './lib/errors/HTTPError.js';
 export * from './lib/errors/RateLimitError.js';
-export * from './lib/RequestManager.js';
 export * from './lib/REST.js';
 export * from './lib/utils/constants.js';
+export * from './lib/utils/types.js';
 export { calculateUserDefaultAvatarIndex, makeURLSearchParams, parseResponse } from './lib/utils/utils.js';
 
 /**
