+++ packages/rest/src/strategies/undiciRequest.ts
@@ -1,7 +1,7 @@
 import { STATUS_CODES } from 'node:http';
 import { URLSearchParams } from 'node:url';
 import { types } from 'node:util';
-import { type RequestInit, request } from 'undici';
+import { type RequestInit, request, Headers } from 'undici';
 import type { ResponseLike } from '../shared.js';
 
 export type RequestOptions = Exclude<Parameters<typeof request>[1], undefined>;
